requirements.txt
selenium==4.15.2
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pdf2image==1.16.3
pytesseract==0.3.10
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
flask==3.0.0
gunicorn==21.2.0
python-docx==1.0.1 